pocket-cfo-parser/requirements.txt
pdfplumber
python-dateutil
pytest
reportlab
pymongo
python-dotenv
google-genai
fastapi
uvicorn
python-multipart
openai
